+++ weblate/checks/models.py
@@ -66,6 +66,7 @@ class WeblateChecksConf(AppConf):
         "weblate.checks.format.PerlFormatCheck",
         "weblate.checks.format.JavaScriptFormatCheck",
         "weblate.checks.format.LuaFormatCheck",
+        "weblate.checks.format.ObjectPascalFormatCheck",
         "weblate.checks.format.SchemeFormatCheck",
         "weblate.checks.format.CSharpFormatCheck",
         "weblate.checks.format.JavaFormatCheck",