+++ docs/changes.rst
@@ -3,6 +3,7 @@ Weblate 4.6
 
 Not yet released.
 
+* Added support for ``object-pascal-format`` used in gettext PO.
 * The auto_translate management command has now a parameter for specifying translation mode.
 * Added support for :ref:`txt`.
 * Added trends and metrics for all objects.